requirements.txt
Flask==2.0.1
Flask-SocketIO==5.1.1
Flask-CORS==3.0.10
python-dotenv==0.19.0
pandas==1.3.3
supabase==1.0.3
gunicorn==20.1.0
eventlet==0.33.3
python-engineio==4.3.1
python-socketio==5.4.0
numpy==1.24.3
scikit-learn==1.3.0
matplotlib==3.7.2
seaborn==0.12.2
bcrypt==4.0.1
Werkzeug==2.3.7
Jinja2==3.1.2
itsdangerous==2.1.2
MarkupSafe==2.1.3
click==8.1.7
python-dateutil==2.8.2
pytz==2023.3
six==1.16.0
Pillow==10.0.0
streamlit==1.29.0
plotly==5.18.0
gevent==23.9.1
gevent-websocket==0.10.1 